backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pymysql
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-dotenv
openai
google-generativeai
